this fixes an issue where having more than one open pull request would cause the generator to create all the PR jobs with the same scm_branch as the first PR job
