+++ .rubocop.yml
@@ -29,6 +29,7 @@ Metrics/LineLength:
     - 'spec/lib/jenkins_pipeline_builder/compiler_spec.rb'
     - 'spec/lib/jenkins_pipeline_builder/view_spec.rb'
     - 'spec/lib/jenkins_pipeline_builder/pull_request_spec.rb'
+    - 'spec/lib/jenkins_pipeline_builder/pull_request_generator_spec.rb'
     - 'lib/jenkins_pipeline_builder/extensions/*'
 
 # Configuration parameters: CountComments.
