+++ lib/jenkins_pipeline_builder.rb
@@ -27,6 +27,7 @@ require 'jenkins_pipeline_builder/version'
 require 'jenkins_pipeline_builder/utils'
 require 'jenkins_pipeline_builder/compiler'
 require 'jenkins_pipeline_builder/module_registry'
+require 'jenkins_pipeline_builder/pull_request_generator'
 require 'jenkins_pipeline_builder/pull_request'
 require 'jenkins_pipeline_builder/view'
 require 'jenkins_pipeline_builder/generator'
